requirements.txt
gradio>=4.44.0
mediapipe>=0.10.14
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.0.0
scikit-learn>=1.4.0
